pyproject.toml
[project]
name = "gradient-mechanics"
version = "0.0.1"
description = "Accelerated deep learning."
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "cvcuda-cu12",
    "numpy>=2.1.3",
    "nvidia-nvimgcodec-cu12>=0.3.0.5",
    "pyav>=13.1.0",
    "pynvvideocodec>=1.0.2",
    "rich>=13.9.4",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "debugpy>=1.8.8",
    "datasets[vision]>=3.1.0",
    "mkdocs-material>=9.5.39",
    "pytest>=8.3.3",
    "tensorboard>=2.18.0",
    "tensorboard-plugin-profile>=2.18.0",
    "torch-tb-profiler>=0.4.3",
]
python-preference = "system"

[tool.uv.sources]
cvcuda-cu12 = { url = "https://github.com/CVCUDA/CV-CUDA/releases/download/v0.11.0-beta/cvcuda_cu12-0.11.0b0-cp311-cp311-linux_x86_64.whl" }
